Recreation of all fragment
